AE-26: Moved MiniCore Routes to (App) Route from (Admin) to let Users access, Ensured Admin Only CRUD in Non MiniCore Tables
