+++ src/routes/(app)/area/+page.svelte
