+++ src/routes/(app)/area/[areaId]/edit/+page.server.ts
@@ -1,38 +1,29 @@
 import { db } from '$lib/server/db';
 import { area } from '$lib/server/db/schema';
-import { requireUser } from '$lib/server/auth/guards';
+import { requireAdmin, requireUser } from '$lib/server/auth/guards';
 import { error, fail, redirect } from '@sveltejs/kit';
 import { eq } from 'drizzle-orm';
 import type { Actions, PageServerLoad } from './$types';
 
-function assertOwnerOrAdmin(user: { id: string; role: string }, item: { createdBy: string }) {
-	if (user.role === 'admin') return;
-	if (item.createdBy !== user.id) throw error(403, 'No autorizado');
-}
-
 export const load: PageServerLoad = async (event) => {
-	const user = requireUser(event);
+	requireUser(event);
 	const { areaId } = event.params;
 	if (!areaId) throw error(400, 'Falta areaId');
 
 	const [item] = await db.select().from(area).where(eq(area.id, areaId));
 	if (!item) throw error(404);
 
-	assertOwnerOrAdmin(user, item);
-
 	return { item, areaId };
 };
 
 export const actions: Actions = {
 	save: async (event) => {
-		const u = requireUser(event);
+		requireAdmin(event);
 		const { areaId } = event.params;
 		if (!areaId) throw error(400, 'Falta areaId');
 
 		const [item] = await db.select().from(area).where(eq(area.id, areaId));
 		if (!item) throw error(404);
-
-		assertOwnerOrAdmin(u, item);
 
 		const data = await event.request.formData();
 		const name = String(data.get('name') ?? '').trim();
@@ -53,14 +44,12 @@ export const actions: Actions = {
 	},
 	//TODO: Change this to SOFT DELETE
 	delete: async (event) => {
-		const u = requireUser(event);
+		requireAdmin(event);
 		const { areaId } = event.params;
 		if (!areaId) throw error(400, 'Falta areaId');
 
 		const [item] = await db.select().from(area).where(eq(area.id, areaId));
 		if (!item) throw error(404);
-
-		assertOwnerOrAdmin(u, item);
 
 		await db.delete(area).where(eq(area.id, areaId));
 		throw redirect(303, '/area');
