+++ src/routes/(app)/area/[areaId]/edit/+page.svelte
