+++ src/routes/(app)/area/[areaId]/sector/+page.server.ts
