+++ src/routes/(app)/area/[areaId]/sector/+page.svelte
