+++ src/routes/(app)/area/[areaId]/sector/[sectorId]/climb/+page.server.ts
