+++ src/routes/(app)/area/[areaId]/sector/[sectorId]/climb/[climbId]/edit/+page.svelte
