+++ src/routes/(app)/area/[areaId]/sector/[sectorId]/climb/[climbId]/grade/+page.server.ts
