+++ src/routes/(app)/area/[areaId]/sector/[sectorId]/climb/[climbId]/grade/[gradeId]/edit/+page.server.ts
