+++ src/routes/(app)/area/[areaId]/sector/[sectorId]/edit/+page.svelte
